requirements.txt
allure-pytest==2.5.2
pytest==3.9.1
pytest-rerunfailures==4.2
requests==2.22.0
vyper-config==0.3.3